pyproject.toml
[project]
name = "DriveMoE"
version = "0.1.0"
description = "Open source codes for Drive-pi0 and DriveMoE"
readme = "README.md"
requires-python = "==3.10.*"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "opencv-python",
    "joblib",
    "bitsandbytes",
    "einops",
    "scikit-learn",
    "gsutil>=5.32",
    "hydra-core",
    "imageio",
    "matplotlib",
    "numpy==1.26.4",
    "omegaconf",
    "pillow",
    "pre-commit>=4.0.1",
    "pretty_errors",
    "protobuf==3.20.3",
    "tensorflow==2.15.0",
    "tensorflow_datasets==4.9.2",
    "torch==2.5.0",
    "torchvision==0.20.0",
    "transformers",
    "tqdm",
    "wandb",
    "ray"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = []

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["A", "B", "E", "F", "I", "RUF", "W"]
ignore = ["E203", "E501", "B006", "B026", "B905"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"]